Semaprate Navigation Bar from App.vue Fixes #48
